show map polygon info on hover
