+++ scripts/actions/helpers/getNodesSelectionAction.js
@@ -1,7 +1,4 @@
-// TODO this doesn't belong to reducers/sankey
-import getSelectedNodesData from 'reducers/helpers/getSelectedNodesData';
-
-export default (nodesIds, visibleNodes, nodesDictWithMeta) => {
+export default (nodesIds, visibleNodes, nodesDictWithMeta, selectedVectorLayers) => {
   if (!nodesIds || !nodesIds[0]) {
     return {
       ids: [],
@@ -11,10 +8,7 @@ export default (nodesIds, visibleNodes, nodesDictWithMeta) => {
     };
   }
 
-  const data = getSelectedNodesData(nodesIds, visibleNodes, nodesDictWithMeta);
-  // filter metas here
-  // data[0].selectedMetas
-
+  const data = getSelectedNodesData(nodesIds, visibleNodes, nodesDictWithMeta, selectedVectorLayers);
   const geoIds = data.map(node => node.geoId).filter(geoId => geoId !== undefined);
   const columnsPos = data.map(node => node.columnPosition);
 
@@ -25,3 +19,29 @@ export default (nodesIds, visibleNodes, nodesDictWithMeta) => {
     columnsPos
   };
 };
+
+const getSelectedNodesData = (selectedNodesIds, visibleNodes, nodesDictWithMeta, selectedVectorLayers) => {
+  if (selectedNodesIds === undefined) {
+    return [];
+  }
+
+  return selectedNodesIds.map(nodeId => {
+    const visibleNode = visibleNodes.find(node => node.id === nodeId);
+    const node = _.cloneDeep(nodesDictWithMeta[nodeId]);
+
+    // add metas from the map layers to the selected nodes data
+    node.selectedMetas = [];
+    let meta;
+    if (selectedVectorLayers.horizontal.uid) {
+      meta = node.meta[selectedVectorLayers.horizontal.uid];
+      if (meta) node.selectedMetas.push(meta);
+    }
+    if (selectedVectorLayers.vertical.uid) {
+      meta = node.meta[selectedVectorLayers.vertical.uid];
+      if (meta) node.selectedMetas.push(meta);
+    }
+
+    return Object.assign(node, visibleNode);
+
+  });
+}
